uniqueIds renaming for consistency, in importVicon; geomtry functions in visualiser moved to separate script.
