+++ python/libraries/bimvee/visualiser.py
@@ -29,6 +29,7 @@ Nothing stops the calling function from applying a color mask to an output in lu
 """
 
 import numpy as np
+from geometry import quat2RotM, rotateUnitVectors, project3dTo2d, slerp, draw_line
 
 # Local imports
 from plotDvsContrast import getEventImageForTimeRange
@@ -116,82 +117,6 @@ class VisualiserFrame(Visualiser):
         y = data['dimY'] if 'dimY' in data else data['frames'][0].shape[0]
         return x, y
 
-def quat2RotM(quat, M=None):
-    if M is None: 
-        M = np.zeros((4, 4))
-    x = quat[0]
-    y = quat[1]
-    z = quat[2]
-    w = quat[3]
-    M[0, 0] = 1 - 2*y**2 - 2*z**2
-    M[0, 1] = 2*x*y - 2*z*w
-    M[0, 2] = 2*x*z + 2*y*w
-    M[1, 0] = 2*x*y + 2*z*w
-    M[1, 1] = 1 - 2*x**2 - 2*z**2
-    M[1, 2] = 2*y*z - 2*x*w
-    M[2, 0] = 2*x*z - 2*y*w
-    M[2, 1] = 2*y*z + 2*x*w
-    M[2, 2] = 1 - 2*x**2 - 2*y**2
-    M[3, 3] = 1
-    return M
-
-def rotateUnitVectors(rotM):
-    xVec = np.expand_dims(np.array([1, 0, 0, 1]), axis=1)
-    yVec = np.expand_dims(np.array([0, 1, 0, 1]), axis=1)
-    zVec = np.expand_dims(np.array([0, 0, 1, 1]), axis=1)
-    allVecs = np.concatenate((xVec, yVec, zVec), axis=1)
-    return rotM.dot(allVecs)
-
-def project3dTo2d(X, Y, Z, smallestRenderDim):
-    projX = X / Z
-    projY = Y / Z
-    projX = (projX + 1) * smallestRenderDim / 2
-    projY = (projY + 1) * smallestRenderDim / 2
-    projX = projX.astype(int)
-    projY = projY.astype(int)
-    return projX, projY
-
-# Spherical linear interpolation, adapted from https://en.wikipedia.org/wiki/Slerp
-DOT_THRESHOLD = 0.9995
-def slerp(q1, q2, time_relative):
-    dot = np.sum(q1 * q2)
-    if dot < 0.0:
-        q2 = -q2
-        dot = -dot
-    if dot > DOT_THRESHOLD:
-        result = q1 + time_relative * (q2 - q1)
-        return (result.T / np.linalg.norm(result, axis=1)).T
-    theta_0 = np.arccos(dot)
-    sin_theta_0 = np.sin(theta_0)
-    theta = theta_0 * time_relative
-    sin_theta = np.sin(theta)
-    s0 = np.cos(theta) - dot * sin_theta / sin_theta_0
-    s1 = sin_theta / sin_theta_0
-    return (s0 * q1) + (s1 * q2)
-
-# adapted from https://stackoverflow.com/questions/50387606/python-draw-line-between-two-coordinates-in-a-matrix
-def draw_line(mat, x0, y0, x1, y1):
-    if not (0 <= x0 < mat.shape[0] and 0 <= x1 < mat.shape[0] and
-            0 <= y0 < mat.shape[1] and 0 <= y1 < mat.shape[1]):
-        print('Invalid coordinates.')
-        return
-    if (x0, y0) == (x1, y1):
-        mat[x0, y0] = 2
-        return
-    # Swap axes if Y slope is smaller than X slope
-    transpose = abs(x1 - x0) < abs(y1 - y0)
-    if transpose:
-        mat = mat.T
-        x0, y0, x1, y1 = y0, x0, y1, x1
-    # Swap line direction to go left-to-right if necessary
-    if x0 > x1:
-        x0, y0, x1, y1 = x1, y1, x0, y0
-    # Compute intermediate coordinates using line equation
-    x = np.arange(x0, x1 + 1)
-    y = np.round(((y1 - y0) / (x1 - x0)) * (x - x0) + y0).astype(x.dtype)
-    # Write intermediate coordinates
-    mat[x, y] = 255
-
 class VisualiserPose6q(Visualiser):
 
     renderX = 300
@@ -203,33 +128,10 @@ class VisualiserPose6q(Visualiser):
         y[-0.5:0.5]
         z[1:2]
     ''' 
-    def centreAndScalePoses(self):
-        data = self.__data
-        poseX = data['pose'][:, 0]
-        poseY = data['pose'][:, 1]
-        poseZ = data['pose'][:, 2]
-        minX = np.min(poseX)
-        maxX = np.max(poseX)
-        minY = np.min(poseY)
-        maxY = np.max(poseY)
-        minZ = np.min(poseZ)
-        maxZ = np.max(poseZ)
-        centreX = (minX + maxX) / 2
-        centreY = (minY + maxY) / 2
-        centreZ = (minZ + maxZ) / 2
-        largestDim = max(maxX-minX, maxY-minY, maxZ-minZ)
-        # Centre the poses
-        poseX = (poseX - centreX) / largestDim
-        poseY = (poseY - centreY) / largestDim
-        poseZ = (poseZ - centreZ) / largestDim + 2
-        data['pose'][:, 0] = poseX
-        data['pose'][:, 1] = poseY
-        data['pose'][:, 2] = poseY
                 
     def set_data(self, data):
         self.__data = data
         self.smallestRenderDim = min(self.renderX, self.renderY)
-        #self.centreAndScalePoses()
         poseX = data['pose'][:, 0]
         poseY = data['pose'][:, 1]
         poseZ = data['pose'][:, 2]
@@ -247,6 +149,7 @@ class VisualiserPose6q(Visualiser):
         poseX = (poseX - centreX) / largestDim
         poseY = (poseY - centreY) / largestDim
         poseZ = (poseZ - centreZ) / largestDim + 2
+        
         self.projX, self.projY = project3dTo2d(poseX, poseY, poseZ, self.smallestRenderDim)
         #
         rotMats = np.apply_along_axis(quat2RotM, axis=1, arr=data['pose'][:, 3:7])
